+++ lib/src/plutus_data/constr_plutus_data.c
@@ -213,7 +213,7 @@ cardano_constr_plutus_data_from_cbor(cardano_cbor_reader_t* reader, cardano_cons
   {
     static const char* validator_name = "constr_plutus_data";
 
-    const cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, CONSTR_PLUTUS_EMBEDDED_GROUP_SIZE);
+    const cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, (uint32_t)CONSTR_PLUTUS_EMBEDDED_GROUP_SIZE);
 
     if (expect_array_result != CARDANO_SUCCESS)
     {
