+++ lib/src/pool_params/single_host_name_relay.c
@@ -149,7 +149,7 @@ cardano_single_host_name_relay_from_cbor(cardano_cbor_reader_t* reader, cardano_
 
   static const char* validator_name = "single_host_name_relay";
 
-  const cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, EMBEDDED_GROUP_SIZE);
+  const cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, (uint32_t)EMBEDDED_GROUP_SIZE);
 
   if (expect_array_result != CARDANO_SUCCESS)
   {
