+++ lib/src/witness_set/bootstrap_witness.c
@@ -162,7 +162,7 @@ cardano_bootstrap_witness_from_cbor(cardano_cbor_reader_t* reader, cardano_boots
 
   static const char* validator_name = "bootstrap_witness";
 
-  cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, EMBEDDED_GROUP_SIZE);
+  cardano_error_t expect_array_result = cardano_cbor_validate_array_of_n_elements(validator_name, reader, (uint32_t)EMBEDDED_GROUP_SIZE);
 
   if (expect_array_result != CARDANO_SUCCESS)
   {
